# docs/openapi-core.yaml
openapi: 3.1.0
info:
  title: Shopping API
  description: API for managing customer, orders and products records along with their orders and products
  version: 1.0.0
  contact:
    name: Shopping Management System
servers:
  - url: http://localhost:4000
...
paths:
...
  /api/public/orders:
    post:
      tags:
        - orders
      summary: Create order
      description: Create a new order
      operationId: createCustomerOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrder'
      responses:
        '201':
          description: Successfully created a new order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: "Could not create order, please check all required fields are correct."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400'
  /api/public/orders/{id}:
    put:
      tags:
        - orders
      summary: Update order
      description: Update an existing order
      operationId: updateCustomerOrder
      parameters:
        - name: id
          in: path
          description: The ID of the order to update
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrder'
      responses:
        '201':
          description: Successfully updated the order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: "Could not update the order, please check all required fields are correct."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400'
        '404':
          description: "No order with that ID was found"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'
  /api/orders:
    get:
      tags:
        - orders
      summary: Get all order
      description: Get all orders in the system
      operationId: getAllOrders
      responses:
        '200':
          description: Successfully retrieved all orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllOrders'
    post:
      tags:
        - orders
      summary: Create order
      description: Create a new order
      operationId: createOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrder'
      responses:
        '201':
          description: Successfully created a new order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: "Could not create order, please check all required fields are correct."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400'
  /api/orders/{id}:
    get:
      tags:
        - orders
      summary: Get order by ID
      description: Get a specific order by their ID
      operationId: getOrderById
      parameters:
        - name: id
          in: path
          description: The ID of the order to retrieve
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successfully retrieved the order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: "No order with that ID was found"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'
    put:
      tags:
        - orders
      summary: Update order
      description: Update an existing order
      operationId: updateOrder
      parameters:
        - name: id
          in: path
          description: The ID of the order to update
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrder'
      responses:
        '201':
          description: Successfully updated the order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: "Could not update the order, please check all required fields are correct."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400'
        '404':
          description: "No order with that ID was found"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'
    delete:
      tags:
        - orders
      summary: Delete order
      description: Delete a order by ID
      operationId: deleteOrder
      parameters:
        - name: id
          in: path
          description: The ID of the order to delete
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successfully deleted the order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: "No order with that ID was found"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'
...
components:
  schemas:
...
    Customer:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "John"
        email:
          type: string
          example: "john.doe@university.edu"
        orderList:
          type: array
          items:
            $ref: '#/components/schemas/Order'
...
    CreateOrder:
      type: object
      required: [customer_id, order_items]
      properties:
        customer_id:
          type: integer
          format: int64
          example: 1
        order_items:
          type: array
          items:
            $ref: '#components/schemas/CreateOrderItem'
    Order:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        created_at:
          type: string
          format: date-time
          example: "2025-09-01T10:15:00.000Z"
        total_amount:
          type: number
          format: float
          example: 3.2
        customer:
          $ref: '#/components/schemas/Customer'
        productList:
          type: array
          items:
            $ref: '#/components/schemas/Product'
...
    AllOrders:
      type: array
      items:
        $ref: '#/components/schemas/Order'
...
    Product:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "John"
        price:
          type: float
          example: 3.2
...
    Error400:
      type: object
      properties:
        message:
          type: string
          example: "bad request"
    Error404:
      type: object
      properties:
        message:
          type: string
          example: "not found"